feat: added postgres to ffi (#1117)
* feat: added postgres to ffi
