+++ crates/cdk-ffi/src/lib.rs
@@ -7,6 +7,8 @@
 pub mod database;
 pub mod error;
 pub mod multi_mint_wallet;
+pub mod postgres;
+pub mod sqlite;
 pub mod token;
 pub mod types;
 pub mod wallet;
